refactor(BikeGoalSummary): To prepare dacc data display
